Update README.md
The evil eye, deeply rooted in various cultures, is believed to ward off negativity and bring good fortune to the wearer. A gold evil eye necklace, with its elegant and timeless design, adds a touch of mystique to any ensemble while serving as a talisman for safeguarding against the malevolent gaze. The delicate craftsmanship and symbolism behind the piece make it a meaningful and stylish accessory, perfect for those who seek both fashion and a touch of spiritual significance in their jewelry choices.
https://beautysalonorbit.com/gold-evil-eye-necklace/